refactor(DocumentSigner): move to Employee/Documents/onboarding/DocumentSigner
Step 2 of the Employee Documents reorganization outlined in buffer.md.
