+++ src/components/Employee/exports/employeeOnboarding.ts
@@ -3,8 +3,8 @@ export { SelfOnboardingFlow } from '../SelfOnboardingFlow/SelfOnboardingFlow'
 export { EmployeeList } from '../EmployeeList/onboarding/EmployeeList'
 export { OnboardingSummary } from '../OnboardingSummary'
 export { Landing } from '../Landing'
-export { DocumentSigner } from '../DocumentSigner'
-export { EmploymentEligibility } from '../DocumentSigner/EmploymentEligibility'
+export { DocumentSigner } from '../Documents/onboarding/DocumentSigner'
+export { EmploymentEligibility } from '../Documents/onboarding/DocumentSigner/EmploymentEligibility'
 
 export { Profile } from '../Profile/onboarding/Profile'
 export { Compensation } from '../Compensation'
