+++ src/components/Employee/index.ts
@@ -8,7 +8,7 @@ export type { FederalTaxesProps } from './FederalTaxes'
 export { StateTaxes } from './StateTaxes'
 export { PaymentMethod } from './PaymentMethod'
 export { Landing } from './Landing'
-export { DocumentSigner } from './DocumentSigner'
+export { DocumentSigner } from './Documents/onboarding/DocumentSigner'
 export { OnboardingFlow } from './OnboardingFlow/OnboardingFlow'
 export { SelfOnboardingFlow } from './SelfOnboardingFlow/SelfOnboardingFlow'
 export { EmployeeDocuments } from './Documents/onboarding/EmployeeDocuments'
@@ -17,8 +17,8 @@ export type { DashboardFlowProps } from './Dashboard'
 export { HomeAddress } from './HomeAddress/management/HomeAddress'
 export type { HomeAddressProps } from './HomeAddress/management/HomeAddress'
 
-export { EmploymentEligibility } from './DocumentSigner/EmploymentEligibility'
-export type { EmploymentEligibilityProps } from './DocumentSigner/EmploymentEligibility'
+export { EmploymentEligibility } from './Documents/onboarding/DocumentSigner/EmploymentEligibility'
+export type { EmploymentEligibilityProps } from './Documents/onboarding/DocumentSigner/EmploymentEligibility'
 
 export { TerminateEmployee } from './Terminations/TerminateEmployee/TerminateEmployee'
 export type { TerminateEmployeeProps } from './Terminations/TerminateEmployee/TerminateEmployee'
