+++ src/index.ts
@@ -420,7 +420,7 @@ export {
   MAX_PREPARERS,
   PREPARER_FIELDS_BY_INDEX,
   preparerFieldName,
-} from '@/components/Employee/DocumentSigner/shared/useSignEmployeeForm'
+} from '@/components/Employee/Documents/onboarding/DocumentSigner/shared/useSignEmployeeForm'
 export type {
   UseSignEmployeeFormProps,
   UseSignEmployeeFormResult,
@@ -441,7 +441,7 @@ export type {
   UsedPreparerFieldProps,
   PreparerTextFieldProps,
   PreparerCheckboxFieldProps,
-} from '@/components/Employee/DocumentSigner/shared/useSignEmployeeForm'
+} from '@/components/Employee/Documents/onboarding/DocumentSigner/shared/useSignEmployeeForm'
 
 // Domain hooks - Company
 export {
